Reorganizando la arquitectura del proyecto:
- Reorganizamos las capas del proyecto con la Meta-Arquitectura "Clean Arquitecture" junto a "MVVM".
- Añadimos las librerias de Retrofit2, Converter-GSON y Corrutinas.
- Añadimos permisos a la aplicación para poder acceder a internet.
- Se creo el archivo RetrofitHelper.kt que contiene la función "getRetrofit()" que al llamarla nos devolverá una instancia de Retrofit preparada para usar.
