+++ app/build.gradle
@@ -50,6 +50,11 @@ dependencies {
     implementation "androidx.fragment:fragment-ktx:1.6.0"
     // Activity
     implementation "androidx.activity:activity-ktx:1.7.2"
+    // Retrofit
+    implementation "com.squareup.retrofit2:retrofit:2.9.0"
+    implementation "com.squareup.retrofit2:converter-gson:2.9.0"
+    //Corrutinas
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.4'
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
