+++ app/src/main/java/com/example/mvvmapp/core/RetrofitHelper.kt
@@ -1,0 +1,13 @@
+package com.example.mvvmapp.core
+
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+object RetrofitHelper {
+    fun getRetrofit(): Retrofit{
+        return Retrofit.Builder()
+            .baseUrl("https://drawsomething-59328-default-rtdb.europe-west1.firebasedatabase.app/")
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+    }
+}
