+++ app/src/main/java/com/example/mvvmapp/data/model/QuoteModel.kt
@@ -1,0 +1,8 @@
+package com.example.mvvmapp.data.model
+
+import com.google.gson.annotations.SerializedName
+
+data class QuoteModel(
+    @SerializedName("quote") val quote: String,
+    @SerializedName("author") val author: String
+)
