--- app/src/main/java/com/example/mvvmapp/model/QuoteModel.kt
@@ -1,3 +1,0 @@
-package com.example.mvvmapp.model
-
-data class QuoteModel(val quote:String, val author:String)
