+++ app/src/main/java/com/example/mvvmapp/ui/view/MainActivity.kt
@@ -1,11 +1,11 @@
-package com.example.mvvmapp.view
+package com.example.mvvmapp.ui.view
 
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Observer
 import com.example.mvvmapp.databinding.ActivityMainBinding
-import com.example.mvvmapp.viewmodel.QuoteViewModel
+import com.example.mvvmapp.ui.viewmodel.QuoteViewModel
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding:ActivityMainBinding
@@ -17,11 +17,16 @@ class MainActivity : AppCompatActivity() {
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
+        // Live data es básicamente el PATRÓN OBSERVE.
+        // Lo que esté dentro de "quoteViewModel.quoteModel.observe", se ejecutará automáticamente cada vez que nuestro objeto con live data sea modificado
         quoteViewModel.quoteModel.observe(this, Observer{ currentQuote ->
             binding.tvQuote.text = currentQuote.quote
             binding.tvAuthor.text = currentQuote.author
         })
+        // --------------------------------------------------------------------
 
+        // El objeto con live data se modifica al llamar a la función randomQuote() del ViewModel.
+        // Para ello añadimos un setOnClickListener al constraintLayout principal. Así cada vez que tocamos la pantalla actualizaremos la cita
         binding.viewContainer.setOnClickListener { quoteViewModel.randomQuote() }
     }
 }
